+++ tests/caseinst/test_getlist.py
@@ -1,0 +1,61 @@
+# -*- coding: utf-8 -*-
+
+import unittest.mock
+
+import pytest
+
+import pycamunda.caseinst
+from tests.mock import raise_requests_exception_mock, not_ok_response_mock
+
+
+def test_getlist_params(engine_url, getlist_input, getlist_output):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url, **getlist_input)
+
+    assert get_instances.url == engine_url + '/case-instance'
+    assert get_instances.query_parameters() == getlist_output
+    assert get_instances.body_parameters() == {}
+
+
+def test_getlist_variable_params(engine_url):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url)
+
+    get_instances.add_variable(name='myvar', value='myval', type_='type', value_info={})
+
+    assert get_instances.query_parameters() == {
+        'variables': {'myvar': {'value': 'myval', 'type': 'type', 'valueInfo': {}}}
+    }
+
+
+@unittest.mock.patch('requests.Session.request')
+def test_getlist_calls_requests(mock, engine_url):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url)
+    get_instances()
+
+    assert mock.called
+    assert mock.call_args[1]['method'].upper() == 'GET'
+
+
+@unittest.mock.patch('requests.Session.request', raise_requests_exception_mock)
+def test_getlist_raises_pycamunda_exception(engine_url):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url)
+    with pytest.raises(pycamunda.PyCamundaException):
+        get_instances()
+
+
+@unittest.mock.patch('requests.Session.request', not_ok_response_mock)
+@unittest.mock.patch('pycamunda.caseinst.CaseInstance', unittest.mock.MagicMock())
+@unittest.mock.patch('pycamunda.base._raise_for_status')
+def test_getlist_raises_for_status(mock, engine_url):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url)
+    get_instances()
+
+    assert mock.called
+
+
+@unittest.mock.patch('requests.Session.request', unittest.mock.MagicMock())
+def test_getlist_returns_case_instances(engine_url):
+    get_instances = pycamunda.caseinst.GetList(url=engine_url)
+    instances = get_instances()
+
+    assert isinstance(instances, tuple)
+    assert all(isinstance(instance, pycamunda.caseinst.CaseInstance) for instance in instances)